Firebase Analytics has been implemented. Some files are changed.
Some part of main has been modified according to the case that after login screen, it was going back to main and then Homepage(), Yet, the routes mean nothing for now. Adjust routes accordingly.
Analytics giving setCurrentScreen errors. Fix that.
Viewing target profile is now available. targetProfile.dart has been created.

+++ lib/classes/customUser.dart
@@ -11,7 +11,7 @@ class customUser
 
   int followers_count;
   int followings_count;
-  int rate;
+  double rate;
 
   //POSTS below
 
@@ -49,7 +49,7 @@ class customUser
 
       followers_count: doc.data()["followers"],
       followings_count: doc.data()["followings"],
-      rate : doc.data()["rate"]
+      rate : doc.data()["rate"].toDouble()
     );
   }
 }
